Merges channel polling and checking services
